general refactoring
