+++ components/Modal/common/styles.ts
@@ -1,0 +1,4 @@
+import * as theme from "../../styles/variables"
+
+export const border = `1px solid ${theme.shades.gray85}`
+export const borderRadius = "0.4rem"
