--- components/Modal/commonStyles.ts
@@ -1,1 +1,0 @@
-export const border = "1px solid #D9D9D9"
