+++ components/Modal/components/Footer.tsx
@@ -1,14 +1,15 @@
 import React, { ReactNode, useContext } from 'react'
 import styled from "styled-components"
-import { border } from "../commonStyles"
-import { StyleContext } from "../common"
+import { border } from "../common/styles"
+import { StyleContext } from "../common/contexts"
 
 const Footer = ({
     children,
     showBorder
   }: {
     children: ReactNode,
-    showBorder: boolean}
+    showBorder: boolean
+  }
 ) => {
   const { padding: {horizontal, vertical} } = useContext(StyleContext)
