+++ components/Modal/components/Header.tsx
@@ -1,15 +1,18 @@
 import React, { ReactNode, useContext } from 'react'
 import styled from 'styled-components'
-import { StyleContext } from "../common"
+import { StyleContext } from "../common/contexts"
+import { borderRadius } from "../common/styles"
 
 interface HeaderProps {
   title: string,
   rightSection?: ReactNode
 }
 
-const borderRadius = "0.4rem"
-
-const Header: React.FC<HeaderProps> = ({ title, rightSection }) => {
+const Header: React.FC<HeaderProps> = (
+  { title, rightSection }
+    :
+  { title: string, rightSection: ReactNode}
+) => {
   const { padding: {horizontal, vertical} } = useContext(StyleContext)
   const Container = styled.div`
     display: flex;
