+++ components/Modal/components/Main.tsx
@@ -1,8 +1,12 @@
-import React, { useContext } from 'react'
+import React, { useContext, ReactNode } from 'react'
 import styled from "styled-components"
-import { StyleContext } from "../common"
+import { StyleContext } from "../common/contexts"
 
-export default ({ children, setBodyRef }) => {
+export default (
+    { children, setBodyRef }
+      :
+    {children: ReactNode, setBodyRef: (instance: HTMLDivElement | null) => void}
+) => {
   const { padding: {horizontal, vertical} } = useContext(StyleContext)
   const Main = styled.div`
     padding: ${`${vertical} ${horizontal} 0rem`};
