+++ components/Modal/variants/interfaces.ts
@@ -1,0 +1,11 @@
+import { ReactNode } from 'react'
+
+export interface ModalProps {
+  header: ReactNode,
+  body: ReactNode,
+  footer: ReactNode,
+}
+
+export interface PanelModalProps extends ModalProps {
+  panelContent: ReactNode
+}
