+++ components/styles/variables.tsx
@@ -82,6 +82,7 @@ export const shades = {
   white: chroma.hsl(0, 0, 1),
   gray95: chroma.hsl(0, 0, 0.95),
   gray90: chroma.hsl(0, 0, 0.9),
+  gray85: chroma.hsl(0, 0, 0.85),
   gray50: chroma.hsl(0, 0, 0.5),
   gray20: chroma.hsl(0, 0, 0.2),
   black: chroma.hsl(0, 0, 0),
